server: Update file upload/delete API

+++ routes/filepond.js
@@ -7,14 +7,12 @@ const textParser = bodyParser.text({type: 'text/plain'})
 
 const uploadPath = './public/upload/'
 
-const storage = multer.diskStorage(
-    {
-        destination: uploadPath,
-        filename: function ( req, file, cb ) {
-            cb(null, file.originalname+ '-' + Date.now()+".png");
-        }
+const storage = multer.diskStorage({
+    destination: uploadPath,
+    filename: function ( req, file, cb ) {
+        cb(null, file.originalname);
     }
-);
+});
 
 let upload = multer({
     storage: storage
@@ -23,7 +21,7 @@ let upload = multer({
 const router = express.Router();
 
 router.post('/img', upload.single('filepond'), (req, res) => {
-    return res.send(req.file.filename)
+    return res.send(req.file)
 });
 
 router.delete('/img', textParser, async (req, res) => {
@@ -31,7 +29,12 @@ router.delete('/img', textParser, async (req, res) => {
         return res.status(404).send("Image Not found.");
     }
 
-    let path = uploadPath + req.body;
+    let item = JSON.parse(req.body)
+    if (!item.filename){
+        return res.status(404).send("Image Not found.");
+    }
+
+    let path = uploadPath + item.filename;
     fs.unlink(path, (err) => {
         if(err){
             console.log(err)
